[IMP] pos_self_order: free qr code ui cleaning
In this commit:
-------------------

- Removed the QR code icon from the desktop view of the `Get QR Code` button.
- Moved the QR code button from the navbar into the burger menu for mobile view.
- Added a new reference option `Free Metal / Wood Stands` in settings.
- Renamed the `Edit` menu item to `Settings`, so it clearly leads the user to
the settings page for the selected config.

task:5003448

X-original-commit: 33fae67fd42cb177997a01bd649ac09a65d0a6c3
