+++ addons/pos_restaurant/static/tests/tours/utils/floor_screen_util.js
@@ -61,7 +61,7 @@ export function clickTableSelectorButton() {
     return [
         {
             content: "click on table selector button",
-            trigger: ".floor-screen .right-buttons button.btn-secondary",
+            trigger: ".floor-screen .right-buttons button i.fa-hashtag",
             run: "click",
         },
     ];
