+++ addons/pos_self_order/__manifest__.py
@@ -49,10 +49,12 @@
             'pos_self_order/static/tests/unit/**/*',
         ],
         'point_of_sale._assets_pos': [
+            'pos_self_order/static/src/backend/qr_order_button/*',
             'pos_self_order/static/src/overrides/**/*',
         ],
         'web.assets_backend': [
             "pos_self_order/static/src/upgrade_selection_field.js",
+            'pos_self_order/static/src/backend/qr_order_button/*',
         ],
         "pos_self_order.assets": [
             "pos_self_order/static/src/app/primary_variables.scss",
