+++ addons/pos_self_order/static/src/overrides/models/pos_store.js
@@ -14,4 +14,14 @@ patch(PosStore.prototype, {
 
         return await super.getServerOrders(...arguments);
     },
+    async redirectToQrForm() {
+        const user_data = await this.data.call("pos.config", "get_pos_qr_order_data", [
+            this.config.id,
+        ]);
+        return await this.action.doAction({
+            type: "ir.actions.client",
+            tag: "pos_qr_stands",
+            params: { data: user_data },
+        });
+    },
 });
