--- addons/pos_self_order/static/src/overrides/screens/floor_screen/floor_screen.js
@@ -1,6 +1,0 @@
-/** @odoo-module **/
-
-import { QrOrderButton } from "@pos_self_order/overrides/components/qr_order_button/qr_order_button";
-import { FloorScreen } from "@pos_restaurant/app/screens/floor_screen/floor_screen";
-
-FloorScreen.components = { ...FloorScreen.components, QrOrderButton };
