+++ addons/pos_self_order/static/src/overrides/screens/floor_screen/floor_screen.xml
@@ -4,7 +4,9 @@
 
     <t t-name="pos_self_order.FloorScreen" t-inherit="pos_restaurant.FloorScreen" t-inherit-mode="extension">
         <xpath expr="//NumpadDropdown" position="before">
-            <QrOrderButton/>
+            <button t-if="!ui.isSmall" class="qr-order-button btn btn-secondary btn-lg" t-on-click="() => this.pos.redirectToQrForm()">
+                Get QR Codes
+            </button>
         </xpath>
     </t>
